# api/openapi.yaml
openapi: 3.0.0
info:
  description: "Demo IPA - middelware API \n Use token
    `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJsb2NhbCI6eyJwYXNzd29yZCI6IiQyYiQxMCRDRzBhWS5aM0hkQ1hXbS5aaDZzYXguSk1HbXpsdjN5c3JqdDdxZS83WGxxYjdSeC5GRzJWdSIsImxvZ2luIjoidmljdG9yQGdlcmZhdWQuY29tIn0sInJvbGUiOiJwcmUiLCJsYW5nIjoiZnIiLCJsZWFndWVzIjpbXSwiX2lkIjoiNWZhZmQxMjY5MzQ2Y2MxZjE1OTBhNDcxIiwibmFtZSI6InZpY3RvckBnZXJmYXVkLmNvbSIsImNyZWF0ZWRfYXQiOiIyMDIwLTExLTE0VDEyOjQ0OjIyLjMwNloiLCJ1cGRhdGVkX2F0IjoiMjAyMC0xMS0xNFQxMjo0NDoyMi4zMDZaIiwiX192IjowLCJpYXQiOjE2MDUzNjU0NTMsImV4cCI6MTYwNjIyOTQ1M30.Vu46o6xuY_njlEmY7v9fukqlHw8Yi72S7kxA9LjB_Yw` to test the authorization filters."
  version: 1.0.0
  title: 'DEMO IPA'
  contact:
    email: 'jm+fujitsu@cheynier.com'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  /users/{userId}:
    get:
      tags:
        - user
      summary: Get user by ID
      description: "Get User based on its ID, private part 'local: { login:<login>, password:<password> }' is dropped from the object"
      operationId: getUser
      parameters:
        - name: userId
          in: path
          description: ID of user to return
          required: true
          schema:
            type: string
      responses:
        "200": 
          $ref: "#/components/responses/UserPublic"
        "401":
          description: Unauthorized          
    put:
      tags:
        - user
      summary: Update an existing user
      description: "Provided fields will be updated"
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          description: ID of user to update
          required: true
          schema:
            type: string
          example: 
            5fa717131562d71c187d9e2b
            
            
        
      requestBody:
        $ref: "#/components/requestBodies/UserUpdate"
      responses:
        "400":
          description: Invalid ID supplied
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "405":
          description: Validation exception 
    delete:
      tags:
        - user
      summary: Delete an existing user
      description: ""
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: ID of user to return
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Invalid ID supplied
        "401":
          description: Unauthorized
        "404":
          description: User not found 
...
servers:
  - url: /api
    description: NFA API - Test Me Live Bitch!
components:
  requestBodies:
...
    UserUpdate:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserPrivate'
          examples: 
            login:
              summary: "Modify users email and name"
              value: {"email": "ichanged@nfa.com", "name": "JavierP"}
  responses:
...
    UserPublic:
      description: "Description"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UserPublic"
...
  schemas:
    UserPublic:
      type: object
      properties:
        _id:
          type: string
        role:
          type: string
          enum: 
            - "user"
            - "pre"
            - "admin"
        username:
          type: string
    UserPrivate:
      type: 'object'
      properties:
        role:
          description: 'Role of the user (Can only be changed with suffcient access rights)'
          type: string
          enum: 
            - "user"
            - "pre"
            - "admin"
        lang: 
          description: 'Preferred Language'
          type: string
        name:
          description: 'Username visible by other players' 
          type: string
        email:
          description: 'Email address' 
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: 'bearer'
      description: 'JWT Token'
      bearerFormat: "JWT"
security:
  - bearerAuth: []